+++ extension/src/NABfunctions.ts
@@ -23,9 +23,11 @@ import * as SettingsLoader from "./Settings/SettingsLoader";
 
 // import { OutputLogger as out } from './Logging';
 
-export async function refreshXlfFilesFromGXlf(): Promise<void> {
+export async function refreshXlfFilesFromGXlf(
+  suppressMessage = false
+): Promise<void> {
   console.log("Running: RefreshXlfFilesFromGXlf");
-  let refreshResult;
+  let refreshResult: LanguageFunctions.RefreshResult;
   try {
     if (XliffEditorPanel.currentPanel?.isActiveTab()) {
       throw new Error(
@@ -37,8 +39,10 @@ export async function refreshXlfFilesFromGXlf(): Promise<void> {
     showErrorAndLog(error);
     return;
   }
-
-  vscode.window.showInformationMessage(getRefreshXlfMessage(refreshResult));
+  const showMessage = suppressMessage ? refreshResult.isChanged() : true;
+  if (showMessage) {
+    vscode.window.showInformationMessage(getRefreshXlfMessage(refreshResult));
+  }
   console.log("Done: RefreshXlfFilesFromGXlf");
 }
 
@@ -173,6 +177,7 @@ export async function findNextUnTranslatedText(
   try {
     const settings = SettingsLoader.getSettings();
     const languageFunctionsSettings = new LanguageFunctionsSettings(settings);
+    // Search active text editor first
     if (vscode.window.activeTextEditor) {
       if (vscode.window.activeTextEditor.document.uri.fsPath.endsWith(".xlf")) {
         foundAnything = await LanguageFunctions.findNextUnTranslatedText(
@@ -184,6 +189,7 @@ export async function findNextUnTranslatedText(
         );
       }
     }
+    // Search any xlf file
     if (!foundAnything) {
       foundAnything = await LanguageFunctions.findNextUnTranslatedText(
         settings,
@@ -193,15 +199,25 @@ export async function findNextUnTranslatedText(
         lowerThanTargetState
       );
     }
+    // Run refresh from g.xlf then search again.
+    if (languageFunctionsSettings.refreshXlfAfterFindNextUntranslated) {
+      if (!foundAnything) {
+        await refreshXlfFilesFromGXlf(true);
+        foundAnything = await LanguageFunctions.findNextUnTranslatedText(
+          settings,
+          SettingsLoader.getAppManifest(),
+          false,
+          languageFunctionsSettings.replaceSelfClosingXlfTags,
+          lowerThanTargetState
+        );
+      }
+    }
   } catch (error) {
     showErrorAndLog(error);
     return;
   }
-
   if (!foundAnything) {
-    vscode.window.showInformationMessage(
-      `No more untranslated texts found. Update XLF files from g.xlf if this was unexpected.`
-    );
+    vscode.window.showInformationMessage(`No more untranslated texts found.`);
   }
   console.log("Done: FindNextUnTranslatedText");
 }
@@ -384,10 +400,8 @@ function getRefreshXlfMessage(changes: RefreshResult): string {
   if (changes.numberOfUpdatedNotes > 0) {
     msg += `${changes.numberOfUpdatedNotes} updated notes, `;
   }
-  if (!isNullOrUndefined(changes.numberOfRemovedNotes)) {
-    if (changes.numberOfRemovedNotes > 0) {
-      msg += `${changes.numberOfRemovedNotes} removed notes, `;
-    }
+  if (changes.numberOfRemovedNotes > 0) {
+    msg += `${changes.numberOfRemovedNotes} removed notes, `;
   }
   if (changes.numberOfUpdatedSources > 0) {
     msg += `${changes.numberOfUpdatedSources} updated sources, `;
@@ -399,6 +413,9 @@ function getRefreshXlfMessage(changes: RefreshResult): string {
     if (changes.numberOfSuggestionsAdded > 0) {
       msg += `${changes.numberOfSuggestionsAdded} added suggestions, `;
     }
+  }
+  if (changes.numberOfReviewsAdded > 0) {
+    msg += `${changes.numberOfReviewsAdded} targets in need of review, `;
   }
   if (msg !== "") {
     msg = msg.substr(0, msg.length - 2); // Remove trailing ,